+++ CHANGELOG.md
@@ -5,6 +5,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3] - 2021-11-10
+### Added
+- Discord channel
+- `load_data` function
+- `save_data_file` function
+- `save_data` method
+### Changed
+- `data` parameter added to GenerativeImage `__init__`
+- `depth` parameter added to `save_image` method
+- `depth` parameter added to `save_fig_file` function
+- `save_image` and `nft_storage` methods background bug fixed
+- `README.md` updated
+- Test system updated
+- `Python 3.10` added to `test.yml`
 ## [0.2] - 2021-10-14
 ### Added
 - `dependabot.yml`
@@ -12,13 +26,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `samila_help` function
 - `test.py`
 - `function_test.py`
+- `overall_test.py`
+- `nft_upload_test.py`
 - `is_same_data` function
 - `save_image` method
 ### Changed
 - `dev-requirements.txt` updated
 - `README.md` updated
 - `__main__.py` updated
-- `test.yml` updated
+- Test system updated
 - `nft_storage` method updated
 ## [0.1] - 2021-09-30
 ### Added
@@ -27,7 +43,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `generate` method
 - `nft_storage` method
 
-[Unreleased]: https://github.com/sepandhaghighi/samila/compare/v0.2...dev
+[Unreleased]: https://github.com/sepandhaghighi/samila/compare/v0.3...dev
+[0.3]: https://github.com/sepandhaghighi/samila/compare/v0.2...v0.3
 [0.2]: https://github.com/sepandhaghighi/samila/compare/v0.1...v0.2
 [0.1]: https://github.com/sepandhaghighi/samila/compare/1058677...v0.1